Heavy refactoring
Include:
- Split core.py up between several files to make it easier to manage.
- Add --version arg to scripts.
- Merge of full & db registries
