+++ requirements.txt
@@ -1,4 +1,3 @@
-coverage==4.0.1
 Flask==0.10.1
 Jinja2==2.8
 MarkupSafe==0.23
@@ -6,8 +5,6 @@ Werkzeug==0.10.4
 demjson==2.2.3
 envoy==0.0.3
 mpld3==0.2
-pytest==2.8.2
-pytest-cov==2.2.0
 python-dateutil==2.4.2
 pytz==2015.4
 requests==2.7.0
