+++ scripts/hadoopinspector_runner.py
@@ -3,25 +3,21 @@
 This source code is protected by the BSD license.  See the file "LICENSE"
 in the source code root directory for the full language or refer to it here:
    http://opensource.org/licenses/BSD-3-Clause
-Copyright 2015 Will Farmer and Ken Farmer
+Copyright 2015, 2016 Will Farmer and Ken Farmer
 """
-
-
-
-import sys
-import os
-import argparse
+import sys, os, argparse
 import logging
 import logging.handlers
-import collections
-import subprocess
-import json
 from pprint import pprint as pp
 from os.path import dirname, basename, isdir, isfile, exists
 from os.path import join as pjoin
 
 sys.path.insert(0, dirname(dirname(os.path.abspath(__file__))))
+from hadoopinspector._version import __version__
 import hadoopinspector.core as core
+import hadoopinspector.registry as registry
+import hadoopinspector.check_runner as check_engine
+import hadoopinspector.check_results as chk_results
 
 runner_logger = None
 
@@ -32,19 +28,19 @@ def main():
     runner_logger = setup_runner_logger(args.log_dir, args.log_level, args.log_to_console)
     runner_logger.info("runner starting now")
 
-    registry = core.Registry()
-    registry.load_registry(args.registry_filename)
-    registry.generate_db_registry(args.table, args.check)
+    reg = registry.Registry()
+    reg.load_registry(args.registry_filename)
+    reg.filter_registry(args.table, args.check)
 
     check_repo = core.CheckRepo(args.check_dir)
-    check_results = core.CheckResults(db_fqfn=args.results_filename)
+    check_results = chk_results.CheckResults(args.instance, args.database, db_fqfn=args.results_filename)
 
-    checker = core.CheckRunner(registry, check_repo, check_results, args.instance, args.database,
-                               args.log_dir, args.log_level)
+    checker = check_engine.CheckRunner(reg, check_repo, check_results, args.instance, args.database,
+                                       args.log_dir, args.log_level)
     checker.add_db_var('hapinsp_instance', args.instance)
     checker.add_db_var('hapinsp_database', args.database)
     checker.add_db_var('hapinsp_ssl',      args.ssl)
-    checker.run_checks_for_tables(args.table)
+    checker.run_checks_for_tables()
 
     if args.report:
         print('')
@@ -110,6 +106,10 @@ def get_args():
     parser.add_argument('--log-level',
                         default='debug',
                         choices=['debug', 'info', 'warning', 'error', 'critical'])
+    parser.add_argument('--version',
+                        action='version',
+                        version=__version__,
+                        help='displays version number')
 
     args = parser.parse_args()
 
@@ -138,29 +138,29 @@ def setup_runner_logger(logdir, log_level, log_to_console):
     log_filename = pjoin(logdir, 'runner.log')
 
     #--- create logger
-    runner_logger = logging.getLogger('RunnerLogger')
-    runner_logger.setLevel(log_level.upper())
+    logger = logging.getLogger('RunnerLogger')
+    logger.setLevel(log_level.upper())
 
     #--- add formatting:
     log_format = '%(asctime)s : %(name)-12s : %(levelname)-8s : %(message)s'
     date_format = '%Y-%m-%d %H.%M.%S'
-    runner_formatter = logging.Formatter(log_format, date_format)
+    formatter = logging.Formatter(log_format, date_format)
 
     #--- create rotating file handler
     file_handler = logging.handlers.RotatingFileHandler(log_filename, maxBytes=1000000, backupCount=20)
-    file_handler.setFormatter(runner_formatter)
-    runner_logger.addHandler(file_handler)
+    file_handler.setFormatter(formatter)
+    logger.addHandler(file_handler)
 
     #--- create console handler:
     if log_to_console:
         console_handler = logging.StreamHandler()
-        console_handler.setFormatter(runner_formatter)
-        runner_logger.addHandler(console_handler)
+        console_handler.setFormatter(formatter)
+        logger.addHandler(console_handler)
 
     #--- ensure any uncaught exceptions get logged:
     sys.excepthook = excepthook
 
-    return runner_logger
+    return logger
 
 
 def excepthook(*args):
